v2

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: pakopian <akopyanpyotrx99@gmail.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/18 20:57:18 by pakopian          #+#    #+#             */
-/*   Updated: 2025/03/23 17:22:36 by pakopian         ###   ########.fr       */
+/*   Updated: 2025/03/26 21:23:21 by pakopian         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,13 +37,13 @@ static void check_modifier(char c, int *result, va_list args)
     else if (c == 's')
         *result += print_string(va_arg(args, char *));
     else if (c == 'p')
-        *result += print_pointer(va_arg(args, size_t));
+        *result += print_pointer(va_arg(args, unsigned long long));
     else if (c == 'i' || c == 'd')
         *result += print_int(va_arg(args, int));
     else if (c == 'u')
         *result += print_unsigned(va_arg(args, unsigned int));
     else if (c == 'x' || c == 'X')
-        *result += print_hex(va_arg(args, size_t), c);
+        *result += print_hex(va_arg(args, unsigned long long), c);
 }
 
 int	ft_printf(const char *str, ...)
